+++ src/vroom_chr.cc
@@ -27,9 +27,14 @@ cpp11::strings read_chr(vroom_vec_info* info) {
       auto str = *b;
       auto val = info->locale->encoder_.makeSEXP(str.begin(), str.end(), true);
       PROTECT(val);
-      if (Rf_xlength(val) < str.end() - str.begin()) {
+      if (std::find(str.begin(), str.end(), '\0') != str.end()) {
         info->errors->add_error(
-            b.index(), col->get_index(), "", "embedded null", b.filename());
+            b.index(),
+            col->get_index(),
+            "",
+            "embedded null",
+            b.filename(),
+            b.line());
       }
 
       SET_STRING_ELT(out, i++, check_na(nas, val));