+++ src/vroom_chr.h
@@ -1,5 +1,6 @@
 #pragma once
 
+#include <algorithm>
 #include <cpp11/strings.hpp>
 
 #include "altrep.h"
@@ -54,10 +55,15 @@ public:
     auto val =
         PROTECT(info.locale->encoder_.makeSEXP(str.begin(), str.end(), true));
 
-    if (Rf_xlength(val) < str.end() - str.begin()) {
-      auto&& itr = info.column->begin();
+    if (std::find(str.begin(), str.end(), '\0') != str.end()) {
+      auto itr = info.column->begin() + i;
       info.errors->add_error(
-          itr.index(), col->get_index(), "", "embedded null", itr.filename());
+          itr.index(),
+          col->get_index(),
+          "",
+          "embedded null",
+          itr.filename(),
+          itr.line());
     }
 
     val = check_na(*info.na, val);